Repurpose `Message` type to also represent produced messages
Currently, `Message` only represents messages that have been delivered
from BlazingMQ; users interact with it through a `MessageHandle`.  In
order to allow users to batch-post multiple `Message`s in one PUT to
the broker, this patch repurposes `Message` to also represent messages
that the producer is constructing, to be sent to along to the broker.
This opens up the `post` function to accept an array of `Message`s
that will be sent in a batch.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>