Add C++ `struct` to represent a subscription
We need a way of passing a description of a consumer subscription from
Python code to the C++ `pybmq` library, which will then convert it to
a `bmqt::Subscription`.  Because want our Cython
layer:(`_ext.{pyi,pyx}`) to be a lean as possible to aid with
debugging, we do not want to expose the `bmqt` types directly into the
Cython layer.  Instead, this patch adds a temporary type
`pybmq::Subscription`, which simply stores the data we will need to
create a `bmqt::Subscription` in `pybmq`.  While this adds an
additional copy to each subscription that is configured, configuring a
subscription itself is a relatively expensive process, subscriptions
are not intended to be frequently modified, and consumers are not
intended to have many subscriptions on each consumer.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>