Retrieve subscription handle from message
In order to expose to a consumer the subscription that caused a message
to be delivered, if any, we need to expose subscription handles to the
user.  In the C++ SDK that we wrap, subscription handles can have many
different types, but in the Python SDK, we only allow subscription
handles to be integers.  Thus, we need some way of associating a message
to an optional integer.

This patch teaches the `pybmq_messageutils` component to retrieve an
integral subscription handle from a BlazingMQ message and return it to
the user as an attribute of a `Message`.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>