Chore: Bump dependency versions
This patch bumps the dependency versions of BDE, ntf-core, and
BlazingMQ/libbmq to the most recent released versions.  ntf-core has
gained support for transport-level compression using zlib, lz4, and
zstd, but since BlazingMQ does not use this and since we do not want a
runtime dependency on the lz4 or zstd libraries, we disable these
while building ntf-core.

Signed-off-by: Patrick M. Niedzielski <pniedzielski@bloomberg.net>